refactor: consolidate Mqlv2IrEmitters helpers; tighten access; relocate stage builders from translator
- Consolidate duplicated helpers: remove AGGREGATE_FUNCTION_NAMES, aggSignature,
  aggColumnSignature, and extractUnnestArrayPath from Mqlv2SelectTranslator; callers
  in the translator now invoke the canonical versions in Mqlv2IrEmitters.
- Relocate stage-construction logic from translator to Mqlv2IrEmitters:
  buildJoinsStage → translateJoins (+ private buildUnnestJoinStage helper),
  buildQueryGroupStageViaIr → translateQueryGroupStage (drops "ViaIr" archaeology suffix).
  Support helpers (isUnnestFunctionTable, irJoinType, collectParentColumnNames,
  columnExpressionOf, and friends) move with them.
- Tighten access on 14 formerly-public static methods in Mqlv2IrEmitters that are
  only called internally: translateArrayLength, translateArrayGet, translateArrayConstructor,
  translateExtract, translateAggregateReference, translateArrayContains,
  translateArrayIntersects, translateExists, translateInSubQuery, translateAnyAllSubquery,
  translateScalarSubquery, translateUnnestExists, translateUnnestScalarSubquery,
  translateInnerQuerySpec.
- aggSignature and AGGREGATE_FUNCTION_NAMES remain public static because
  buildAggNames / collectHavingOnlyAggs in the translator (a different package) still
  call them; those callers will move in a subsequent cleanup.
