+++ src/main/java/com/mongodb/hibernate/internal/translate/Mqlv2SelectTranslator.java
@@ -24,11 +24,7 @@ import com.mongodb.hibernate.internal.FeatureNotSupportedException;
 import com.mongodb.hibernate.internal.translate.mqlv2.Mqlv2IrEmitters;
 import com.mongodb.hibernate.internal.translate.mqlv2.Mqlv2TranslationContext;
 import com.mongodb.mqlv2.Serializer;
-import com.mongodb.mqlv2.ast.BinaryOpType;
-import com.mongodb.mqlv2.ast.Expr;
-import com.mongodb.mqlv2.ast.JoinType;
 import com.mongodb.mqlv2.ast.Stage;
-import com.mongodb.mqlv2.ast.Value;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -46,11 +42,7 @@ import org.hibernate.persister.internal.SqlFragmentPredicate;
 import org.hibernate.query.spi.QueryOptions;
 import org.hibernate.query.sqm.SetOperator;
 import org.hibernate.query.sqm.function.SelfRenderingFunctionSqlAstExpression;
-import org.hibernate.query.sqm.sql.internal.BasicValuedPathInterpretation;
-import org.hibernate.query.sqm.sql.internal.EntityValuedPathInterpretation;
-import org.hibernate.query.sqm.sql.internal.SqmParameterInterpretation;
 import org.hibernate.sql.ast.Clause;
-import org.hibernate.sql.ast.SqlAstJoinType;
 import org.hibernate.sql.ast.SqlAstNodeRenderingMode;
 import org.hibernate.sql.ast.SqlAstTranslator;
 import org.hibernate.sql.ast.tree.SqlAstNode;
@@ -92,7 +84,6 @@ import org.hibernate.sql.ast.tree.from.NamedTableReference;
 import org.hibernate.sql.ast.tree.from.QueryPartTableReference;
 import org.hibernate.sql.ast.tree.from.TableGroup;
 import org.hibernate.sql.ast.tree.from.TableGroupJoin;
-import org.hibernate.sql.ast.tree.from.TableReference;
 import org.hibernate.sql.ast.tree.from.TableReferenceJoin;
 import org.hibernate.sql.ast.tree.from.ValuesTableReference;
 import org.hibernate.sql.ast.tree.insert.InsertSelectStatement;
@@ -151,7 +142,7 @@ final class Mqlv2SelectTranslator implements SqlAstTranslator<JdbcSelect> {
     private boolean hasJoins = false;
     /**
      * Maps join-alias → array field path for unnest joins (struct arrays). Populated by
-     * {@link #buildUnnestJoinStage} when emitting an unwind stage; consulted by
+     * {@link Mqlv2IrEmitters#translateJoins} when emitting an unwind stage; consulted by
      * {@link Mqlv2IrEmitters#translateExpression} (via the shared map in {@link Mqlv2TranslationContext})
      * so that column references qualified by the unnest alias resolve to {@code <arrayPath>.<column>}.
      */
@@ -298,13 +289,14 @@ final class Mqlv2SelectTranslator implements SqlAstTranslator<JdbcSelect> {
         // Only count non-unnest joins: unnest joins translate to | unwind and don't introduce
         // a new MQLv2 alias, so they must not trigger the qualifier-prefix path in column-ref rendering.
         this.hasJoins = root.getTableGroupJoins().stream()
-                .anyMatch(tgj -> !isUnnestFunctionTable(tgj.getJoinedGroup().getPrimaryTableReference()));
+                .anyMatch(tgj -> !(tgj.getJoinedGroup().getPrimaryTableReference() instanceof FunctionTableReference ftr
+                        && "unnest".equals(ftr.getFunctionExpression().getFunctionName())));
 
         var ntr = (NamedTableReference) root.getPrimaryTableReference();
         var outerQualifiers = collectOuterQualifiers(querySpec);
         var ctx = newContext().withSubquerySupport(outerQualifiers, () -> correlatedVarCounter++);
         var stage = Mqlv2IrEmitters.translateFromStage(ntr, hasJoins);
-        stage = buildJoinsStage(stage, root, querySpec, ctx);
+        stage = Mqlv2IrEmitters.translateJoins(stage, root, querySpec, ctx);
         stage = Mqlv2IrEmitters.translateMatch(stage, querySpec, ctx);
 
         List<@Nullable String> aggNames = null;
@@ -355,246 +347,8 @@ final class Mqlv2SelectTranslator implements SqlAstTranslator<JdbcSelect> {
                 .toList();
 
         var subStages = translations.stream().map(SpecTranslation::stage).toList();
-        Stage groupStage = buildQueryGroupStageViaIr(operator, subStages);
+        Stage groupStage = Mqlv2IrEmitters.translateQueryGroupStage(queryGroup, subStages, () -> correlatedVarCounter++);
         return new SpecTranslation(serializer.serialize(groupStage), translations.get(0).fieldNames(), groupStage);
-    }
-
-    /**
-     * Builds the driver-mqlv2 {@link Stage} for a UNION/UNION_ALL/INTERSECT/EXCEPT query group. Called when all
-     * sub-specs were translated via the IR path (each sub-spec's {@link SpecTranslation#stage()} is non-null).
-     *
-     * <ul>
-     *   <li>UNION_ALL: {@code from <<(sub1), (sub2), ...>> | unwind $*}
-     *   <li>UNION: {@code from <<(sub1), (sub2), ...>> | unwind $* | distinct}
-     *   <li>INTERSECT: {@code left | match (count(let $__vN = $ in (right | match ($ == $__vN))) > 0)}
-     *   <li>EXCEPT: same shape but with {@code == 0}
-     * </ul>
-     */
-    private Stage buildQueryGroupStageViaIr(SetOperator operator, List<@Nullable Stage> subStages) {
-        return switch (operator) {
-            case UNION_ALL, UNION -> {
-                var subPipeExprs = subStages.stream()
-                        .map(s -> (Expr) new Expr.SubPipelineExpr(s))
-                        .toList();
-                Stage source = new Stage.FromStageSimple(new Expr.BagConstructor(subPipeExprs));
-                Stage unwind = new Stage.UnwindSimpleStage(source, new Expr.UnwindExpr(new Expr.CurrentValue()));
-                yield operator == SetOperator.UNION ? new Stage.DistinctStage(unwind) : unwind;
-            }
-            case INTERSECT, EXCEPT -> {
-                var leftStage = subStages.get(0);
-                var rightStage = subStages.get(1);
-                // right | match ($ == $__vN)
-                var varName = "__v" + correlatedVarCounter++;
-                Stage rightWithMatch = new Stage.MatchStage(
-                        rightStage,
-                        new Expr.BinaryOp(BinaryOpType.EQ, new Expr.CurrentValue(), new Expr.VarRef(varName)));
-                // let $__vN = $ in (right | match ...)
-                Expr letExpr = new Expr.LetExpr(
-                        List.of(Map.entry(varName, new Expr.CurrentValue())),
-                        new Expr.SubPipelineExpr(rightWithMatch));
-                // count(let ...) > 0  or  == 0
-                Expr countExpr = new Expr.FunctionCall("count", List.of(letExpr));
-                Expr predicate = new Expr.BinaryOp(
-                        operator == SetOperator.INTERSECT ? BinaryOpType.GT : BinaryOpType.EQ,
-                        countExpr,
-                        new Expr.ValueLit(new Value.VInt(0)));
-                yield new Stage.MatchStage(leftStage, predicate);
-            }
-            default -> throw new FeatureNotSupportedException("Unsupported set operator: " + operator);
-        };
-    }
-
-    // ---- Join translation ----
-
-    /**
-     * Traverses the table-group-join list and builds a chain of {@link Stage} nodes
-     * ({@link Stage.UnwindComplexStage} for unnest joins, {@link Stage.JoinStage} for regular joins).
-     *
-     * <p>Populates {@code unnestAliasToFieldPath} (via the shared map in {@code ctx}) so that subsequent
-     * column-reference rendering resolves unnest aliases correctly.
-     */
-    private Stage buildJoinsStage(Stage prev, TableGroup root, QuerySpec querySpec, Mqlv2TranslationContext ctx) {
-        var s = prev;
-        for (var tgj : root.getTableGroupJoins()) {
-            var joinedGroup = tgj.getJoinedGroup();
-            var primaryRef = joinedGroup.getPrimaryTableReference();
-            if (isUnnestFunctionTable(primaryRef)) {
-                var rootAlias = ((NamedTableReference) root.getPrimaryTableReference()).getIdentificationVariable();
-                s = buildUnnestJoinStage(s, tgj, (FunctionTableReference) primaryRef, querySpec, rootAlias);
-            } else {
-                var joinNtr = (NamedTableReference) primaryRef;
-                var joinCollName = joinNtr.getTableExpression();
-                var joinAlias = joinNtr.getIdentificationVariable();
-                var joinPredicate = tgj.getPredicate();
-                if (joinPredicate == null) {
-                    throw new FeatureNotSupportedException("Join without ON condition is not supported in MQLv2");
-                }
-                var condExpr = Mqlv2IrEmitters.translatePredicate(joinPredicate, ctx);
-                s = new Stage.JoinStage(
-                        s, irJoinType(tgj.getJoinType()), joinAlias, new Expr.VarRef(joinCollName), condExpr);
-                s = buildJoinsStage(s, joinedGroup, querySpec, ctx);
-            }
-        }
-        return s;
-    }
-
-    /**
-     * Builds a {@link Stage.UnwindComplexStage} that explodes the array field and preserves all parent-entity columns
-     * in the unwind body.
-     *
-     * <p>Also records the alias→array-field-path mapping in {@link #unnestAliasToFieldPath} so that subsequent
-     * column-reference rendering resolves {@code alias.column} to {@code arrayFieldPath.column}.
-     */
-    private Stage buildUnnestJoinStage(
-            Stage prev,
-            TableGroupJoin tgj,
-            FunctionTableReference ftr,
-            QuerySpec querySpec,
-            @Nullable String rootAlias) {
-        var arrayPath = extractUnnestArrayPath(ftr);
-        var alias = extractUnnestAlias(tgj.getJoinedGroup());
-        // Unwind targets the field on the current (outer) document without qualifier prefix.
-        var arrayFieldPath = columnExpressionOf(arrayPath);
-        var internalVarName = "__elem";
-
-        // Collect all parent-entity column names referenced anywhere in the QuerySpec.
-        // These must be enumerated in the unwind body to be preserved across the stage.
-        var parentCols = new LinkedHashSet<String>();
-        collectParentColumnNames(querySpec, rootAlias, parentCols);
-        // Always include the array field itself: subsequent pipeline stages reference it as
-        // arrayField.subField (e.g., lineItems.sku), so the unwind body must carry it forward.
-        parentCols.add(arrayFieldPath);
-
-        // Build the body document: {col1: col1, ..., arrayField: $__elem}
-        var fields = new ArrayList<Map.Entry<Expr, Expr>>();
-        for (var col : parentCols) {
-            Expr keyExpr = new Expr.ValueLit(new Value.VString(col));
-            Expr valExpr;
-            if (col.equals(arrayFieldPath)) {
-                // Map the array field to the element variable so subsequent match stages work.
-                valExpr = new Expr.VarRef(internalVarName);
-            } else {
-                valExpr = new Expr.FieldAccess(new Expr.CurrentValue(), col);
-            }
-            fields.add(Map.entry(keyExpr, valExpr));
-        }
-        var bodyExpr = new Expr.DocumentConstructor(fields);
-
-        // Source expression: the array field on the current document.
-        var sourceExpr = new Expr.FieldAccess(new Expr.CurrentValue(), arrayFieldPath);
-
-        // Register the alias so downstream column-reference rendering resolves it correctly.
-        unnestAliasToFieldPath.put(alias, arrayFieldPath);
-
-        return new Stage.UnwindComplexStage(prev, internalVarName, sourceExpr, bodyExpr);
-    }
-
-    /**
-     * Maps a Hibernate {@link SqlAstJoinType} to the driver-mqlv2 {@link JoinType}. Returns {@code null} for INNER
-     * joins because {@link Stage.JoinStage} treats a {@code null} join type as an inner join (and the Serializer omits
-     * the keyword).
-     */
-    private static @Nullable JoinType irJoinType(SqlAstJoinType joinType) {
-        return switch (joinType) {
-            case INNER -> null;
-            case LEFT -> JoinType.LEFT_OUTER;
-            case RIGHT -> JoinType.RIGHT_OUTER;
-            case FULL -> JoinType.FULL_OUTER;
-            default -> throw new FeatureNotSupportedException("Unsupported join type: " + joinType);
-        };
-    }
-
-    /**
-     * Collects all parent-entity column names referenced in the given {@code QuerySpec}. A column is considered
-     * "parent" if its qualifier matches {@code rootAlias} (or is null/empty). This drives the body object inside the
-     * complex unwind stage so that parent fields are preserved.
-     *
-     * <p>Scans SELECT, WHERE, ORDER BY, GROUP BY, and HAVING clauses.
-     */
-    private static void collectParentColumnNames(QuerySpec querySpec, @Nullable String rootAlias, Set<String> result) {
-        // SELECT
-        for (var sel : querySpec.getSelectClause().getSqlSelections()) {
-            collectParentColsFromExpr(sel.getExpression(), rootAlias, result);
-        }
-        // WHERE
-        var where = querySpec.getWhereClauseRestrictions();
-        if (where != null && !where.isEmpty()) {
-            collectParentColsFromPredicate(where, rootAlias, result);
-        }
-        // ORDER BY
-        var sortSpecs = querySpec.getSortSpecifications();
-        if (sortSpecs != null) {
-            for (var sort : sortSpecs) {
-                collectParentColsFromExpr(sort.getSortExpression(), rootAlias, result);
-            }
-        }
-        // GROUP BY
-        for (var expr : querySpec.getGroupByClauseExpressions()) {
-            collectParentColsFromExpr(expr, rootAlias, result);
-        }
-        // HAVING
-        var having = querySpec.getHavingClauseRestrictions();
-        if (having != null && !having.isEmpty()) {
-            collectParentColsFromPredicate(having, rootAlias, result);
-        }
-    }
-
-    private static void collectParentColsFromPredicate(
-            Predicate predicate, @Nullable String rootAlias, Set<String> result) {
-        if (predicate instanceof ComparisonPredicate cp) {
-            collectParentColsFromExpr(cp.getLeftHandExpression(), rootAlias, result);
-            collectParentColsFromExpr(cp.getRightHandExpression(), rootAlias, result);
-        } else if (predicate instanceof Junction junction) {
-            for (var p : junction.getPredicates()) collectParentColsFromPredicate(p, rootAlias, result);
-        } else if (predicate instanceof GroupedPredicate gp) {
-            collectParentColsFromPredicate(gp.getSubPredicate(), rootAlias, result);
-        } else if (predicate instanceof NegatedPredicate np) {
-            collectParentColsFromPredicate(np.getPredicate(), rootAlias, result);
-        } else if (predicate instanceof NullnessPredicate np) {
-            collectParentColsFromExpr(np.getExpression(), rootAlias, result);
-        } else if (predicate instanceof BooleanExpressionPredicate bp) {
-            collectParentColsFromExpr(bp.getExpression(), rootAlias, result);
-        } else if (predicate instanceof InListPredicate ilp) {
-            collectParentColsFromExpr(ilp.getTestExpression(), rootAlias, result);
-            for (var e : ilp.getListExpressions()) collectParentColsFromExpr(e, rootAlias, result);
-        }
-        // Other predicate types (EXISTS, IN subquery, etc.) are not walked — their subqueries
-        // are separate query scopes and don't need to contribute columns to the unwind body.
-    }
-
-    private static void collectParentColsFromExpr(Expression expr, @Nullable String rootAlias, Set<String> result) {
-        if (expr instanceof ColumnReference cr) {
-            var q = cr.getQualifier();
-            if (q == null || q.isEmpty() || q.equals(rootAlias)) {
-                result.add(cr.getColumnExpression());
-            }
-        } else if (expr instanceof BasicValuedPathInterpretation<?> bvpi) {
-            collectParentColsFromExpr(bvpi.getColumnReference(), rootAlias, result);
-        } else if (expr instanceof BinaryArithmeticExpression bae) {
-            collectParentColsFromExpr(bae.getLeftHandOperand(), rootAlias, result);
-            collectParentColsFromExpr(bae.getRightHandOperand(), rootAlias, result);
-        } else if (expr instanceof SelfRenderingFunctionSqlAstExpression<?> fn) {
-            for (var arg : fn.getArguments()) {
-                if (arg instanceof Expression e) collectParentColsFromExpr(e, rootAlias, result);
-            }
-        } else if (expr instanceof SqmParameterInterpretation spi) {
-            collectParentColsFromExpr(spi.getResolvedExpression(), rootAlias, result);
-        }
-        // Literals, parameters, subqueries — no column references to collect.
-    }
-
-    /**
-     * Extracts the simple column expression from a path expression. Used to render unwind targets without qualifier
-     * prefix.
-     */
-    private static String columnExpressionOf(Expression expr) {
-        if (expr instanceof ColumnReference cr) {
-            return cr.getColumnExpression();
-        } else if (expr instanceof BasicValuedPathInterpretation<?> bvpi) {
-            return bvpi.getColumnReference().getColumnExpression();
-        }
-        throw new FeatureNotSupportedException("Expected simple column reference for unwind path; got: "
-                + expr.getClass().getSimpleName());
     }
 
     private Mqlv2TranslationContext newContext() {
@@ -629,7 +383,9 @@ final class Mqlv2SelectTranslator implements SqlAstTranslator<JdbcSelect> {
             if (isAggregateFunction(sel.getExpression())) {
                 var name = "_agg" + aggIdx++;
                 aggSignatureToName.put(
-                        aggSignature((SelfRenderingFunctionSqlAstExpression<?>) sel.getExpression()), name);
+                        Mqlv2IrEmitters.aggSignature(
+                                (SelfRenderingFunctionSqlAstExpression<?>) sel.getExpression(), hasJoins),
+                        name);
                 result.add(name);
             } else {
                 result.add(null);
@@ -669,7 +425,7 @@ final class Mqlv2SelectTranslator implements SqlAstTranslator<JdbcSelect> {
             Expression expr, Map<String, SelfRenderingFunctionSqlAstExpression<?>> result) {
         if (isAggregateFunction(expr)) {
             var fn = (SelfRenderingFunctionSqlAstExpression<?>) expr;
-            var sig = aggSignature(fn);
+            var sig = Mqlv2IrEmitters.aggSignature(fn, hasJoins);
             if (!aggSignatureToName.containsKey(sig)) {
                 var name = "_agg" + aggSignatureToName.size();
                 aggSignatureToName.put(sig, name);
@@ -681,38 +437,9 @@ final class Mqlv2SelectTranslator implements SqlAstTranslator<JdbcSelect> {
         }
     }
 
-    private String aggSignature(SelfRenderingFunctionSqlAstExpression<?> fn) {
-        var args = fn.getArguments();
-        if (args.isEmpty() || args.get(0) instanceof Star || args.get(0) instanceof EntityValuedPathInterpretation<?>) {
-            return fn.getFunctionName() + ":*";
-        }
-        if (args.get(0) instanceof Expression argExpr) {
-            return fn.getFunctionName() + ":" + aggColumnSignature(argExpr);
-        }
-        return fn.getFunctionName() + ":?";
-    }
-
-    // Includes the table qualifier (when present) so that count(a.id) and count(b.id)
-    // in a joined query produce distinct signatures.
-    private String aggColumnSignature(Expression expr) {
-        if (expr instanceof ColumnReference cr) {
-            if (hasJoins && cr.getQualifier() != null && !cr.getQualifier().isEmpty()) {
-                return cr.getQualifier() + "." + cr.getColumnExpression();
-            }
-            return cr.getColumnExpression();
-        } else if (expr instanceof BasicValuedPathInterpretation<?> bvpi) {
-            return aggColumnSignature(bvpi.getColumnReference());
-        } else {
-            throw new FeatureNotSupportedException("Expected simple column reference in aggregate; got: "
-                    + expr.getClass().getSimpleName());
-        }
-    }
-
-    private static final Set<String> AGGREGATE_FUNCTION_NAMES = Set.of("count", "sum", "avg", "min", "max");
-
     private static boolean isAggregateFunction(Expression expr) {
         return expr instanceof SelfRenderingFunctionSqlAstExpression<?> fn
-                && AGGREGATE_FUNCTION_NAMES.contains(fn.getFunctionName());
+                && Mqlv2IrEmitters.AGGREGATE_FUNCTION_NAMES.contains(fn.getFunctionName());
     }
 
     private static boolean selectHasAggregates(SelectClause selectClause) {
@@ -736,43 +463,6 @@ final class Mqlv2SelectTranslator implements SqlAstTranslator<JdbcSelect> {
         for (var tgj : group.getTableGroupJoins()) {
             collectAffectedTableNamesRecursive(tgj.getJoinedGroup(), names);
         }
-    }
-
-    // ---- Unnest helpers ----
-
-    /**
-     * @return true iff the table reference is a {@link FunctionTableReference} whose function descriptor identifies as
-     *     "unnest".
-     */
-    private static boolean isUnnestFunctionTable(TableReference ref) {
-        return ref instanceof FunctionTableReference ftr
-                && "unnest".equals(ftr.getFunctionExpression().getFunctionName());
-    }
-
-    /**
-     * Returns the single argument expression to {@code unnest(<arg>)}. Throws {@link FeatureNotSupportedException} if
-     * the argument is not a simple {@link ColumnReference} or {@link BasicValuedPathInterpretation} (which would mean
-     * the user passed a literal array, a function call, or some other non-path expression).
-     */
-    private static Expression extractUnnestArrayPath(FunctionTableReference ftr) {
-        var args = ftr.getFunctionExpression().getArguments();
-        if (args.size() != 1) {
-            throw new FeatureNotSupportedException("unnest() requires exactly one argument; got " + args.size());
-        }
-        var arg = args.get(0);
-        if (arg instanceof ColumnReference || arg instanceof BasicValuedPathInterpretation<?>) {
-            return (Expression) arg;
-        }
-        throw new FeatureNotSupportedException("unnest() argument must be a path expression on an outer entity; got: "
-                + arg.getClass().getSimpleName());
-    }
-
-    /**
-     * Returns the identification variable that aliases the rows of the unnest output (the "a" in {@code FROM o.array
-     * a}).
-     */
-    private static String extractUnnestAlias(TableGroup group) {
-        return ((FunctionTableReference) group.getPrimaryTableReference()).getIdentificationVariable();
     }
 
     // SqlAstWalker visitor methods — only the ones used above are implemented;
